backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.4.0
torchvision==0.19.0
transformers==4.44.2
numpy==1.26.4
scikit-learn==1.5.1
faiss-cpu==1.8.0  # Use faiss-gpu if you have a GPU
pillow==10.4.0